# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TCG Collection Manager API
  description: REST API for managing trading card game collections, including games, sets, cards, user collections, decks, and booster packs.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    AdminApiKey:
      type: http
      scheme: bearer
      description: Admin API key for protected endpoints

    SessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie for authenticated user endpoints
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Game not found"
...
    CreateCollectionRequest:
      type: object
      required:
        - game_id
        - name
      properties:
        game_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "My Pokemon Collection"
    CollectionResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
    UpdateCollectionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "My Renamed Collection"
    CollectionListItem:
      type: object
      required:
        - id
        - game_id
        - name
        - created_at
        - game_name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        game_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "My Pokemon Collection"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        game_name:
          type: string
          example: "Pokemon TCG"
        game_image_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/pokemon-logo.png"
    CollectionDetail:
      type: object
      required:
        - id
        - game_id
        - name
        - created_at
        - game_name
        - card_count
      properties:
        id:
          type: integer
          format: int64
          example: 1
        game_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "My Pokemon Collection"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        game_name:
          type: string
          example: "Pokemon TCG"
        game_image_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/pokemon-logo.png"
        card_count:
          type: integer
          format: int64
          example: 150
    CollectionCardItem:
      type: object
      required:
        - id
        - name
        - collector_number
        - attributes
        - set_id
        - set_name
        - quantity
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Pikachu"
        collector_number:
          type: string
          example: "025"
        image_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/pikachu.png"
        attributes:
          type: object
          additionalProperties:
            type: string
          example:
            rarity: "common"
            type: "electric"
        set_id:
          type: integer
          format: int64
          example: 1
        set_name:
          type: string
          example: "Shrouded Fable"
        quantity:
          type: integer
          format: int64
          example: 3
    CardQuantityUpdate:
      type: object
      required:
        - card_id
        - quantity
      properties:
        card_id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          description: The new quantity (0 to remove from collection)
          example: 3

    # Deck schemas
...
paths:
...
  /api/collections:
    get:
      summary: List user's collections
      description: Get all collections owned by the authenticated user
      operationId: listCollections
      tags:
        - Collections
      security:
        - SessionCookie: []
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionListItem'
        '401':
          description: Unauthorized - no valid session
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a collection
      description: Create a new card collection for a specific game
      operationId: createCollection
      tags:
        - Collections
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '200':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '401':
          description: Unauthorized - no valid session
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Game not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/collections/{collection_id}:
    get:
      summary: Get collection details
      description: Get details of a specific collection
      operationId: getCollection
      tags:
        - Collections
      security:
        - SessionCookie: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The collection ID
          example: 1
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Collection not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update collection
      description: Update the name of a collection
      operationId: updateCollection
      tags:
        - Collections
      security:
        - SessionCookie: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The collection ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionRequest'
      responses:
        '204':
          description: Collection updated successfully
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Collection not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete collection
      description: Delete a collection and all associated cards and decks
      operationId: deleteCollection
      tags:
        - Collections
      security:
        - SessionCookie: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The collection ID
          example: 1
      responses:
        '204':
          description: Collection deleted successfully
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Collection not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/collections/{collection_id}/cards:
    get:
      summary: List cards in collection
      description: Get all cards in a specific collection with quantities
      operationId: listCollectionCards
      tags:
        - Collections
      security:
        - SessionCookie: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The collection ID
          example: 1
      responses:
        '200':
          description: List of cards in collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionCardItem'
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Collection not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update card quantities
      description: Update the quantities of cards in a collection. Set quantity to 0 to remove a card.
      operationId: updateCollectionCards
      tags:
        - Collections
      security:
        - SessionCookie: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The collection ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CardQuantityUpdate'
            example:
              - card_id: 1
                quantity: 3
              - card_id: 5
                quantity: 0
      responses:
        '204':
          description: Cards updated successfully
        '400':
          description: Bad request - card doesn't belong to collection's game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Card does not belong to collection's game"
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection or card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                collectionNotFound:
                  value:
                    error: "Collection not found"
                cardNotFound:
                  value:
                    error: "Card not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...